requirements.txt
easydict==1.9
matplotlib==3.5.1
natsort==8.0.2
numpy==1.21.2
opencv_python==4.5.4.60
Pillow==9.0.1
PyYAML==6.0
scikit_image==0.19.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.10.1+cu113
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.11.2+cu113
tqdm==4.62.3
yacs==0.1.8
